requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.3.5,<2.0.0
pydantic==2.4.2
spacy==3.7.2
